Remove 'resourceful?' check.
This method requires that all resources should implement a 'to_#{format}' API
when this is actually responsability of the renderer that Rails will use for
the response format. This is particually annoying when implementing a new format
that is based in an existing serialization protocol like using a ':hal' format
while using 'to_json' to serialize the resources.