Test with released Rails 4.2 stable branch
The released 4.2.0 contains a bug that does not correctly verify the
response status code to determine whether to include the body content
and remove content headers in case of head requests like 204 no content.

It is not a big deal and it's unlikely to affect many people, but running
responders tests against 4.2.0 fails some tests returning a body with a space
" ", where we expect an empty one "", and the stable branch already
contains the necessary fixes:

https://github.com/rails/rails/commit/bc1760fe646e29a6de909ed639eb6c45dca88111
https://github.com/rails/rails/commit/2910956fc9006d3c26f56992dab630fa97b7f7bc

This does not affect the use of responders with 4.2.0.