Respond with 422 for non-GET HTML requests with errors
When responding to a non-GET HTML request that had errors on the
resource, responders would not set any status, which means Rails would
default to `200 OK`. This has been long working like this but also a
long discussion in responders to change that behavior, to be more inline
with how other types of requests (like JSON/XML) handle responses by
setting the HTTP status to 422 Unprocessable Entity when responding with
an error.

More recently with the Turbo/Hotwire library, they've added support to
render form responses as errors with 4xx statuses [1], so it makes even
more sense for responders to make the move here.

This change now makes responders use the 422 Unprocessable Entity status
in such cases when the resource has errors, and should work more out of
the box with Turbo/Hotwire.

Please note that this is a possible breaking change if you're relying on
the previous status to trigger any behavior for errors in your app.

[1] https://github.com/hotwired/turbo/pull/39