Make error status for HTML/JS responses fully configurable
This essentially makes the previous change merged via #223 configurable,
so that we can keep the current behavior on existing apps, returning
`200 OK` even for errors, while changing the generator for new apps to
be configured with `422 Unprocessable Entity`, making them work better
with Hotwire/Turbo installations out of the box.

Please note that these defaults are likely to be swapped in a future
major version of Responders.